Register Clang 16.0 as the default toolchain
Add Clang 16.0 as the default C++ toolchain. Both GCC 12.2 and Clang
16.0 are tested in the Github check workflow. The test workflow also
verifies that sanitizer features `asan`, `tsan`, and `ubsan` are
available with Clang.

The creation of the huffman tree contained an expression without a well
defined order of evaluation, resulting in different behavior between GCC
and Clang. This is updated to use a well defined order and reflected in
the tests.

Also support building on MacOS, though not with sanitizers.

Change-Id: I2c02446a5e419fdc2e7b6af23a43c002f2343f08