Add clang-tidy configuration
Pull in @bazel_clang_tidy to enable clang-tidy with a hermetic
toolchain. This job also adds a clang-tidy job to the check workflow.

To run clang-tidy on source files:

 bazel build //... --config=clang-tidy

For clang-tidy checks that have fixes, those can be applied with

 bazel run //:clang-tidy-fix

This commit also applies fixes and check suppressions to source files.

In order to allow different rulesets for test code, tests have been
moved into a separate test directory.

Change-Id: If22bbd78ae36068d9425898723db5480127e2ea2