construct huffman table from symbol spans
Define class template `symbol_span` that represents an inclusive range
from first to last. The `symbol_bitsize` constructor of `huffman::table`
now takes a range of pair-likes of `symbol_span` and bitsize, allowin
gthe same bitsize to be associated with a contiguous range of symbols.

For example, the static Huffman tree corresponding to the literal/length
alphabet in section 3.2.6 of the DEFLATE RFC can be defined as:

 constexpr auto table =
  huffman::table<std::uint16_t, 288>{
    huffman::symbol_bitsize,
    {{{  0, 143}, 8},
     {{144, 255}, 9},
     {{256, 279}, 7},
     {{280, 287}, 8}}};

resolves #91

Change-Id: I50f7ea63561de0bb785c85467cdb943829b65edf
Co-Authored-By: Gary Miguel <garymm@garymm.org>