always use DEFLATE canonical form for huffman::table codes
Update huffman::table constructors to always call canonicalize(). As
canonicalize() is now called in the constructor, this method is now
private.

Note that the table_contents constructor is an exception and does not
call canonicalize().

File decode_test.cpp does not yet update the Huffman table to use
canonical form as this requires the recalculating the encoded bits.

Change-Id: I5597dc51ea5e9f086c71e2f9396fccd5ace923b2