add `boost::ut::range_eq`
Define the function `boost::ut::range_eq` that compares elements in two
ranges. If the elements are not equal, this function displays the first
mismatch in the two ranges or if one range is larger than another.

This commit also updates `boost-ut` to allow use of ranges in
parameterized tests.

resolves #106

Change-Id: I24b4f37927799c7eb3b68d3ac2ce159f60e3d94f