test data for fixed and dynamic huffman trees
and a script to generate it

I guess the nice way to do this would be to bring in rules_python and
generate the data in a genrule, or pull in zlib and have our test
generate the compressed data at run-time. We can add that later if
we find keeping the compressed and decompressed data in sync is annoying
enough.

Change-Id: Iaba4043956a2eb6fd0f51aef40414ae948a7311e